+++ yampt.editor/source/dialog/settings/editor_settings_dialog.hpp
@@ -9,8 +9,10 @@ class QPushButton;
 class QStackedWidget;
 class settings_store_t;
 class appearance_settings_view_t;
+class cleaning_settings_view_t;
 class editor_paths_view_t;
 class merge_settings_view_t;
+class sub_record_rules_view_t;
 
 class editor_settings_dialog_t : public QDialog
 {
@@ -33,6 +35,8 @@ private:
 	QPushButton * m_apply_button = nullptr;
 
 	appearance_settings_view_t * m_appearance_view = nullptr;
+	cleaning_settings_view_t * m_cleaning_view = nullptr;
 	editor_paths_view_t * m_paths_view = nullptr;
 	merge_settings_view_t * m_merge_view = nullptr;
+	sub_record_rules_view_t * m_sub_record_rules_view = nullptr;
 };