+++ yampt.editor/source/editor_window.cpp
@@ -134,11 +134,10 @@ void editor_window_t::setup_toolbar()
 	connect(
 	    merge_action, &QAction::triggered, m_plugin_workspace_view, &plugin_workspace_view_t::on_create_merged_patch);
 
-	// WIP: Clean All button hidden until ITM detection issues are resolved
-	// auto * clean_action = new QAction(tr("Clean All"), this);
-	// clean_action->setToolTip(tr("Remove ITMs, evil GMSTs, and junk cells from all plugins"));
-	// toolbar->addAction(clean_action);
-	// connect(clean_action, &QAction::triggered, m_plugin_workspace_view, &plugin_workspace_view_t::on_clean_all);
+	auto * clean_action = new QAction(tr("Clean All"), this);
+	clean_action->setToolTip(tr("Remove evil GMSTs and junk cells from all plugins"));
+	toolbar->addAction(clean_action);
+	connect(clean_action, &QAction::triggered, m_plugin_workspace_view, &plugin_workspace_view_t::on_clean_all);
 
 	toolbar->addSeparator();
 
@@ -178,5 +177,6 @@ void editor_window_t::closeEvent(QCloseEvent * event)
 void editor_window_t::on_open_settings()
 {
 	editor_settings_dialog_t dialog(m_settings, this);
-	dialog.exec();
+	if (dialog.exec() == QDialog::Accepted)
+		m_plugin_workspace_view->on_settings_changed();
 }